% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildData_T1DM.R
\name{buildData_T1}
\alias{buildData_T1}
\title{Build a Covariate Data for cohort}
\usage{
buildData_T1(cdm_bbdd, cdm_schema, results_sc, cohortTable, acohortId = 1)
}
\arguments{
\item{cdm_bbdd}{A connection for a OMOP database via DatabaseConnector}

\item{cdm_schema}{A name for OMOP schema}

\item{results_sc}{A name for result schema}

\item{cohortTable}{A name of the result cohort}

\item{acohortId}{A Cohort number}
}
\value{
A covariateData object
}
\description{
Build a Covariate Data for cohort
}
\examples{
# Not yet
}
