% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildData.R
\name{buildFollowUp}
\alias{buildFollowUp}
\title{Build a Follow-up Data for cohort}
\usage{
buildFollowUp(
  cdm_bbdd,
  cdm_schema,
  results_sc,
  cohortTable,
  acohortId = 1,
  bbdd_covar
)
}
\arguments{
\item{cdm_bbdd}{A connection for a OMOP database via DatabaseConnector}

\item{cdm_schema}{A name for OMOP schema}

\item{results_sc}{A name for result schema}

\item{cohortTable}{A name of the result cohort}

\item{acohortId}{A Cohort number}

\item{bbdd_covar}{A data.table create by transformToFlat function}
}
\value{
A data.frame object
}
\description{
Build a Follow-up Data for cohort
}
\examples{
# Not yet
}
