% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_Eunomia.R
\name{connect_Eunomia}
\alias{connect_Eunomia}
\title{Eunomia Connection}
\usage{
connect_Eunomia()
}
\value{
SQL connection
}
\description{
Make a SQL connection to Eunomia database.
}
\examples{
connection <- connect_Eunomia()
DatabaseConnector::querySql(connection, "SELECT COUNT(*) FROM person;")
DatabaseConnector::getTableNames(connection, databaseSchema = 'main')
}
