% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creacio_synpuf.R
\name{crea_synpuf}
\alias{crea_synpuf}
\title{SYNPUF Creation as SQL}
\usage{
crea_synpuf()
}
\value{
SQL connection
}
\description{
Build a SYNPUF database as SQL
}
\examples{
connection <- crea_synpuf()
DatabaseConnector::querySql(connection,
         "SELECT COUNT(*) FROM person;")
DatabaseConnector::getTableNames(connection,
                                 databaseSchema = 'main')
}
