% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildData.R
\name{createC10_TimeCovariateSettings}
\alias{createC10_TimeCovariateSettings}
\title{Auxiliar function to create Time form C10 diagnosis}
\usage{
createC10_TimeCovariateSettings(useC10_Time = TRUE)
}
\arguments{
\item{useC10_Time}{Logical valor}
}
\value{
covariateSettings object
}
\description{
Auxiliar function to create Time form C10 diagnosis
}
\examples{
#Not yet
}
