% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildData.R
\name{createHTNRx_TimeCovariateSettings}
\alias{createHTNRx_TimeCovariateSettings}
\title{Auxiliar function to create Time form HTN medication diagnosis}
\usage{
createHTNRx_TimeCovariateSettings(useHTNRx_Time = TRUE)
}
\arguments{
\item{useHTNRx_Time}{Logical valor}
}
\value{
covariateSettings object
}
\description{
Auxiliar function to create Time form HTN medication diagnosis
}
\examples{
#Not yet
}
