% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildData.R
\name{createSmokingCovariateSettings}
\alias{createSmokingCovariateSettings}
\title{Auxiliar function to create Smoking status}
\usage{
createSmokingCovariateSettings(useSmoking = TRUE)
}
\arguments{
\item{useSmoking}{Logical valor}
}
\value{
covariateSettings object
}
\description{
Auxiliar function to create Smoking status
}
\examples{
#Not yet
}
