% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildData.R
\name{createT1DM_AgeCovariateSettings}
\alias{createT1DM_AgeCovariateSettings}
\title{Auxiliar function to create Age for first T1DM diagnosis}
\usage{
createT1DM_AgeCovariateSettings(useT1DM_Age = TRUE)
}
\arguments{
\item{useT1DM_Age}{Logical valor}
}
\value{
covariateSettings object
}
\description{
Auxiliar function to create Age for first T1DM diagnosis
}
\examples{
#Not yet
}
