% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildData.R
\name{createT1Rx_AgeCovariateSettings}
\alias{createT1Rx_AgeCovariateSettings}
\title{Auxiliar function to create Age for first T1Rx drug}
\usage{
createT1Rx_AgeCovariateSettings(useT1Rx_Age = TRUE)
}
\arguments{
\item{useT1Rx_Age}{Logical valor}
}
\value{
covariateSettings object
}
\description{
Auxiliar function to create Age for first T1Rx drug
}
\examples{
#Not yet
}
