% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildData.R
\name{createT1Rx_TimeCovariateSettings}
\alias{createT1Rx_TimeCovariateSettings}
\title{Auxiliar function to create Time form T1DM diagnosis}
\usage{
createT1Rx_TimeCovariateSettings(useT1Rx_Time = TRUE)
}
\arguments{
\item{useT1Rx_Time}{Logical valor}
}
\value{
covariateSettings object
}
\description{
Auxiliar function to create Time form T1DM diagnosis
}
\examples{
#Not yet
}
