% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildData.R
\name{createT2DM_TimeCovariateSettings}
\alias{createT2DM_TimeCovariateSettings}
\title{Auxiliar function to create Time form T2DM diagnosis}
\usage{
createT2DM_TimeCovariateSettings(useT2DM_Time = TRUE)
}
\arguments{
\item{useT2DM_Time}{Logical valor}
}
\value{
covariateSettings object
}
\description{
Auxiliar function to create Time form T2DM diagnosis
}
\examples{
#Not yet
}
