% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildData.R
\name{FunCovar}
\alias{FunCovar}
\title{Funcion para pasar del servidor a una tabla plana
Necesita totes les coses del servidor més quina cohort volem contruir.}
\usage{
FunCovar(cdm_bbdd, cdm_schema, results_sc, cohortTable, acohortId)
}
\arguments{
\item{cdm_bbdd}{A connection for a OMOP database via DatabaseConnector}

\item{cdm_schema}{A name for OMOP schema}

\item{results_sc}{A name for result schema}

\item{cohortTable}{A name of the result cohort}

\item{acohortId}{A Cohort number}
}
\value{
List with three objects: descriptive for numeric and category and database ready to analysis.
}
\description{
Funcion para pasar del servidor a una tabla plana
Necesita totes les coses del servidor més quina cohort volem contruir.
}
\examples{
#Not yet
}
