% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildData.R
\name{getDbuseT1Rx_TimeCovariateData}
\alias{getDbuseT1Rx_TimeCovariateData}
\title{Auxiliar function to create Time from T1Rx diagnosis status SQL implementation}
\usage{
getDbuseT1Rx_TimeCovariateData(
  connection,
  oracleTempSchema = NULL,
  cdmDatabaseSchema,
  cohortTable = "#cohort_person",
  cohortId = -1,
  cdmVersion = "5",
  rowIdField = "subject_id",
  covariateSettings,
  aggregated = FALSE
)
}
\arguments{
\item{connection}{A connection for a OMOP database via DatabaseConnector}

\item{oracleTempSchema}{Only for Oracle Database}

\item{cdmDatabaseSchema}{A name for OMOP schema}

\item{cohortTable}{A name of the result cohort}

\item{cohortId}{A Cohort number}

\item{cdmVersion}{CDM version}

\item{rowIdField}{Column with the subject identification}

\item{covariateSettings}{covariateSettings object generetad via FeatureExtraction}

\item{aggregated}{Logical value}
}
\value{
Function to use with FeatureExtraction to build covariateData
}
\description{
Auxiliar function to create Time from T1Rx diagnosis status SQL implementation
}
\examples{
#Not yet
}
