% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStudy.R
\name{runStudy}
\alias{runStudy}
\title{Title}
\usage{
runStudy(
  cdm_bbdd,
  cdm_schema,
  results_sc,
  cohortTable,
  rundiag = FALSE,
  exportFolder
)
}
\arguments{
\item{cdm_bbdd}{A connection for a OMOP database via DatabaseConnector}

\item{cdm_schema}{A name for OMOP schema}

\item{results_sc}{A name for result schema}

\item{cohortTable}{A name of the result cohort}

\item{rundiag}{Logical.}

\item{exportFolder}{Directory to save Diagnostic}
}
\value{

}
\description{
Title
}
\examples{
# Not yet
}
