% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildData.R
\name{transformToFlat}
\alias{transformToFlat}
\title{Transform covariateData object into FlatTable}
\usage{
transformToFlat(covariateData)
}
\arguments{
\item{covariateData}{A covariateDate object}
}
\value{
A data.table with the covariate data
}
\description{
Transform covariateData object into FlatTable
}
\examples{
#Not yet
}
