% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildData_T1DM.R
\name{transformToFlat_T1}
\alias{transformToFlat_T1}
\title{Transform covariateData object into FlatTable}
\usage{
transformToFlat_T1(covariateData, covariateData_temp)
}
\arguments{
\item{covariateData}{A covariateDate object}

\item{covariateData_temp}{A Temporal covariateDate object}
}
\value{
A data.table with the covariate data
}
\description{
Si volem agafar l'últim valor podem canviar-ho
}
\examples{
#Not yet
}
